+++ src/main/java/it/smartcommunitylab/challenges/app/configuration/StandardGroupSettings.java
@@ -17,8 +17,8 @@ public class StandardGroupSettings {
 	private String model;
 	private List<String> modeConcepts;
 	private int minLevel;
-	private Map<String, Integer> modeMax;
-	private Map<String, Integer> modeMin;
+	private Map<String, Double> modeMax;
+	private Map<String, Double> modeMin;
 
 	public Date getStart() {
 		return start;
@@ -44,19 +44,19 @@ public class StandardGroupSettings {
 		this.modeConcepts = modeConcepts;
 	}
 
-	public Map<String, Integer> getModeMax() {
+	public Map<String, Double> getModeMax() {
 		return modeMax;
 	}
 
-	public void setModeMax(Map<String, Integer> modeMax) {
+	public void setModeMax(Map<String, Double> modeMax) {
 		this.modeMax = modeMax;
 	}
 	
-	public Map<String, Integer> getModeMin() {
+	public Map<String, Double> getModeMin() {
 		return modeMin;
 	}
 
-	public void setModeMin(Map<String, Integer> modeMin) {
+	public void setModeMin(Map<String, Double> modeMin) {
 		this.modeMin = modeMin;
 	}
